Remove `include/grpcpp/impl/codegen/time.h`
