+++ include/grpcpp/impl/codegen/time.h
@@ -19,73 +19,9 @@
 #ifndef GRPCPP_IMPL_CODEGEN_TIME_H
 #define GRPCPP_IMPL_CODEGEN_TIME_H
 
-// IWYU pragma: private, include <grpcpp/support/time.h>
+// IWYU pragma: private
 
-#include <chrono>
-
-#include <grpc/impl/codegen/grpc_types.h>
-#include <grpcpp/impl/codegen/config.h>
-
-namespace grpc {
-
-/** If you are trying to use CompletionQueue::AsyncNext with a time class that
-    isn't either gpr_timespec or std::chrono::system_clock::time_point, you
-    will most likely be looking at this comment as your compiler will have
-    fired an error below. In order to fix this issue, you have two potential
-    solutions:
-
-      1. Use gpr_timespec or std::chrono::system_clock::time_point instead
-      2. Specialize the TimePoint class with whichever time class that you
-         want to use here. See below for two examples of how to do this.
- */
-template <typename T>
-class TimePoint {
- public:
-  // If you see the error with methods below, you may need either
-  // i) using the existing types having a conversion class such as
-  // gpr_timespec and std::chrono::system_clock::time_point or
-  // ii) writing a new TimePoint<YourType> to address your case.
-  TimePoint(const T& /*time*/) = delete;
-  gpr_timespec raw_time() = delete;
-};
-
-template <>
-class TimePoint<gpr_timespec> {
- public:
-  /* NOLINTNEXTLINE(google-explicit-constructor) */
-  TimePoint(const gpr_timespec& time) : time_(time) {}
-  gpr_timespec raw_time() { return time_; }
-
- private:
-  gpr_timespec time_;
-};
-
-}  // namespace grpc
-
-namespace grpc {
-
-// from and to should be absolute time.
-void Timepoint2Timespec(const std::chrono::system_clock::time_point& from,
-                        gpr_timespec* to);
-void TimepointHR2Timespec(
-    const std::chrono::high_resolution_clock::time_point& from,
-    gpr_timespec* to);
-
-std::chrono::system_clock::time_point Timespec2Timepoint(gpr_timespec t);
-
-template <>
-class TimePoint<std::chrono::system_clock::time_point> {
- public:
-  /* NOLINTNEXTLINE(google-explicit-constructor) */
-  TimePoint(const std::chrono::system_clock::time_point& time) {
-    Timepoint2Timespec(time, &time_);
-  }
-  gpr_timespec raw_time() const { return time_; }
-
- private:
-  gpr_timespec time_;
-};
-
-}  // namespace grpc
+/// TODO(chengyuc): Remove this file after solving compatibility.
+#include <grpcpp/support/time.h>
 
 #endif  // GRPCPP_IMPL_CODEGEN_TIME_H
